pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "This project contains a web application 'Task manager'"
authors = ["Valerii-Denisov <klerk919191@gmail.com>"]
packages = [{ include = "task_manager"}]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Licence :: Free for non-commercial use",
    "Operating System :: POSIX",
    "Programming Language :: Python",
]
repository = "https://github.com/Valerii-Denisov/python-project-52/"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.4"
python-dotenv = "^0.21.0"
gunicorn = "^20.1.0"
dj-database-url = "^0.5.0"
psycopg2-binary = "^2.9.5"
whitenoise = "^6.2.0"
django-bootstrap4 = "^22.3"
django-filter = "^22.1"
rollbar = "^0.16.3"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
coverage = "^7.0.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
